arm7: split SND into SND_global and SND_channel
