+++ arm7/lib/include/SND_global.h
@@ -1,5 +1,5 @@
-#ifndef GUARD_SND_H
-#define GUARD_SND_H
+#ifndef GUARD_SND_GLOBAL_H
+#define GUARD_SND_GLOBAL_H
 
 #include "nitro/types.h"
 
@@ -8,8 +8,7 @@ void SND_Disable(void);
 void SND_Shutdown(void);
 void SND_BeginSleep(void);
 void SND_EndSleep(void);
-void SND_SetMasterPan(int pan);
 void SND_SetMasterVolume(int vol);
 void SND_SetOutputSelector(int leftOutputFrom, int rightOutputFrom, int outputCh1ToMixer, int outputCh3ToMixer);
 
-#endif //GUARD_SND_H
+#endif //GUARD_SND_GLOBAL_H
