+++ arm7/lib/include/SND_main.h
@@ -12,7 +12,5 @@ u32 SND_WaitForIntervalTimer(void);
 void SND_SendWakeupMessage(void);
 void SNDi_LockMutex(void);
 void SNDi_UnlockMutex(void);
-void SND_SetMasterVolume(int vol);
-void SND_SetMasterPan(int pan);
 
 #endif //GUARD_SND_MAIN_H
