+++ arm7/lib/src/SND_command.c
@@ -7,10 +7,11 @@
 #include "PXI_fifo.h"
 #include "OS_message.h"
 #include "OS_system.h"
-#include "SND.h"
+#include "SND_channel.h"
 #include "SND_alarm.h"
 #include "SND_capture.h"
 #include "SND_exChannel.h"
+#include "SND_global.h"
 #include "SND_main.h"
 #include "SND_seq.h"
 #include "SND_work.h"
