+++ arm7/lib/src/SND_work.c
@@ -1,5 +1,6 @@
 #include "SND_work.h"
 
+#include "SND_channel.h"
 #include "SND_exChannel.h"
 #include "SND_capture.h"
